Android: Apply Error Prone suggested fixes for a some checks (part 3)
Checks:
  * EmptyTopLevelDeclaration
  * ExtendsObject
  * IgnoredPureGetter
  * InlineTrivialConstant
  * LongDoubleConversion
  * NarrowCalculation
  * NullablePrimitive
  * OperatorPrecedence
  * ReturnAtTheEndOfVoidFunction
  * ReturnFromVoid
  * SuperCallToObjectMethod
  * UnnecessaryLongToIntConversion
  * UnnecessaryParentheses
  * UnnecessaryStringBuilder
  * VoidUsed

Bug: 40661145
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5891417
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Owners-Override: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1360930}

+++ components/paint_preview/player/android/junit/src/org/chromium/components/paintpreview/player/PlayerUserFrustrationDetectorTest.java
@@ -32,7 +32,7 @@ public class PlayerUserFrustrationDetectorTest {
                 0,
                 detectionCallback.getCallCount());
         for (int i = 0; i < tapsCount; i++) {
-            detector.recordUnconsumedTap(startTime + i * tapsWindow);
+            detector.recordUnconsumedTap(startTime + i * ((long) tapsWindow));
         }
         Assert.assertEquals(
                 "Frustration callback should have been called once",
@@ -54,7 +54,7 @@ public class PlayerUserFrustrationDetectorTest {
             startTime += (tapsCount + 2) * tapsWindow;
             for (int j = 0; j < tapsCount; j++) {
                 int delay = i == j ? 10 : 0;
-                detector.recordUnconsumedTap(startTime + j * tapsWindow + delay);
+                detector.recordUnconsumedTap(startTime + j * ((long) tapsWindow) + delay);
             }
             Assert.assertEquals(
                     "Frustration callback shouldn't have been called",
@@ -65,7 +65,7 @@ public class PlayerUserFrustrationDetectorTest {
         // Perform a valid tap sequence.
         startTime += (tapsCount + 2) * tapsWindow;
         for (int i = 0; i < tapsCount; i++) {
-            detector.recordUnconsumedTap(startTime + i * tapsWindow);
+            detector.recordUnconsumedTap(startTime + i * ((long) tapsWindow));
         }
         Assert.assertEquals(
                 "Frustration callback should have been called once",
